+++ completions/spin.zsh
@@ -17,8 +17,8 @@ _spin() {
     _arguments "${_arguments_options[@]}" \
 '-c+[Path to a custom configuration file]:CONFIG_PATH: ' \
 '--config-path=[Path to a custom configuration file]:CONFIG_PATH: ' \
-'-l+[Listens to stdout of the provided command and traps interrupt events from reaching the command]:TAIL_COMMAND: ' \
-'--listen-command=[Listens to stdout of the provided command and traps interrupt events from reaching the command]:TAIL_COMMAND: ' \
+'(-f --follow)-l+[Continuously listens to the stdout of the provided command and prevents interrupt events (Ctrl + C) from reaching the command]:LISTEN_COMMAND: ' \
+'(-f --follow)--follow-command=[Continuously listens to the stdout of the provided command and prevents interrupt events (Ctrl + C) from reaching the command]:LISTEN_COMMAND: ' \
 '--z-generate-shell-completions=[Print completions to stdout]:GENERATE_SHELL_COMPLETIONS: ' \
 '-f[Follow (tail) the contents of the file]' \
 '--follow[Follow (tail) the contents of the file]' \